result.go is json.go and codetype.go
