+++ types/json.go
@@ -7,40 +7,6 @@ import (
 	"github.com/gogo/protobuf/jsonpb"
 )
 
-const (
-	CodeTypeOK uint32 = 0
-)
-
-// IsOK returns true if Code is OK.
-func (r ResponseCheckTx) IsOK() bool {
-	return r.Code == CodeTypeOK
-}
-
-// IsErr returns true if Code is something other than OK.
-func (r ResponseCheckTx) IsErr() bool {
-	return r.Code != CodeTypeOK
-}
-
-// IsOK returns true if Code is OK.
-func (r ResponseDeliverTx) IsOK() bool {
-	return r.Code == CodeTypeOK
-}
-
-// IsErr returns true if Code is something other than OK.
-func (r ResponseDeliverTx) IsErr() bool {
-	return r.Code != CodeTypeOK
-}
-
-// IsOK returns true if Code is OK.
-func (r ResponseQuery) IsOK() bool {
-	return r.Code == CodeTypeOK
-}
-
-// IsErr returns true if Code is something other than OK.
-func (r ResponseQuery) IsErr() bool {
-	return r.Code != CodeTypeOK
-}
-
 //---------------------------------------------------------------------------
 // override JSON marshalling so we dont emit defaults (ie. disable omitempty)
 // note we need Unmarshal functions too because protobuf had the bright idea
